+++ filter_classified_reads/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Peter Kruczkiewicz"""
 __email__ = 'peter.kruczkiewicz@gmail.com'
-__version__ = '0.1.0'
+__version__ = '0.2.0'