+++ HISTORY.rst
@@ -2,6 +2,14 @@
 History
 =======
 
+0.2.0 (2019-09-23)
+------------------
+
+* Use ``seqtk subseq`` instead of screed for pulling reads of interest from files
+* Added external dependencies for ``seqtk`` for pulling reads from input FASTQs and ``pbgzip`` for parallel block Gzip for output of Gzipped FASTQs
+* Removed ``screed`` Python dependency
+
+
 0.1.0 (2019-04-15)
 ------------------